Updates projects recipe to use project hash instead of tuple
Previously to specify projects to clone in the attributes you had to put
entries in an array in a specific order.  This was confusing, ill-documented
and buggy. This change will allow for future additions to the projects
configuration

Now the projects attribute is an array of hashes with keys like:
- name
- url
- workspace_path

- This also adds the ability to specify where you wanted the project installed
  both from the users home directory and from /
- Adds examples in soloist rc to allow for integration testing multiple project
  configuration variations

This is the first step to resolving issue #5
[Finishes #62184140]