Fix hook_name string replacement
sub replaces first occurrence only, gsub replaces all occurences.
RuboCop complained that we should use tr instead of gsub. We didn't
fight it.

We believe that using the same code in the implementation and tests is a
bad idea. The tests were passing because the code that copied had the
same mistake (i.e. sub instead of gsub/tr).

Signed-off-by: Gerhard Lazu <glazu@pivotal.io>